refactor: extract SearchEvents, FileEvents, SubtitleEvents from MediaLive.Show
Complete the event extraction by moving the three largest event groups
into dedicated modules, including their handle_async dispatches.

show.ex is now 559 LOC (down from 2,081), containing only mount,
handle_params, thin one-liner dispatches for all 66 handle_event
clauses, handle_info (PubSub), handle_async dispatches, and one
private helper.

New modules:
- SearchEvents (~310 LOC): manual/auto search, download_from_search
- FileEvents (~480 LOC): metadata refresh, rescan, file CRUD, rename
- SubtitleEvents (~140 LOC): subtitle search/download/delete
